requirements.txt
faker
pandas
flask
python-dotenv
sqlalchemy
pymysql
cryptography